# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VoiceTokenResponse": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "identity": {
            "title": "Identity",
            "type": "string"
          },
          "room": {
            "title": "Room",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "ws_url": {
            "title": "Ws Url",
            "type": "string"
          }
        },
        "required": [
          "token",
          "ws_url",
          "room",
          "identity",
          "expires_in"
        ],
        "title": "VoiceTokenResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Openvoice API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/channels/{channel_id}/voice-token": {
      "post": {
        "description": "Authorized voice-channel join (ADR-0005): requires membership and the\nCONNECT_VOICE capability on this specific channel; SPEAK controls whether\nthe token can publish. Room name and identity are derived server-side.",
        "operationId": "issue_channel_voice_token_api_v1_channels__channel_id__voice_token_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Channel Voice Token",
        "tags": [
          "voice"
        ]
      }
    },
...
  }
}